+++ server/src/cvServer.py
@@ -995,6 +995,10 @@ class Data:
         # Add an action listener to the video feed window.
         cv2.setMouseCallback(self.vWindow, self.mouseClicked, None)
 
+        # Add an initial default area of interest that encompasses the entire screen.
+        # Although we set them to 240 by 180, they turn out to be 320 by 240.
+        self.addAreaOfInterest(3, 317, 3, 237)
+
         # The method set is used to set different properties of the VideoCapture instance.
         # 3 determines that the width should be set; 4, the height.
         # (3 is CV_CAP_PROP_FRAME_WIDTH, and 4 is CV_CAP_PROP_FRAME_HEIGHT, but the aliases don't work)